Move test fixtures to the parent package
